Chỉnh sửa hơi bị nhiều
Chuyển các mảng dữ liệu như specialties, services, hospital sang các file Data.js riêng lẻ trong thư mục Data.

Đã lấy được thông tin và hiển thị ModalService khi nhấn vào.
