+++ src/containers/HomePage/Data/serviceData.js
@@ -1,0 +1,14 @@
+const services = [
+  { img: "/icon-service/1.webp", id: "service.specialist" },
+  { img: "/icon-service/2.webp", id: "service.remote" },
+  { img: "/icon-service/3.webp", id: "service.general" },
+  { img: "/icon-service/4.webp", id: "service.lab" },
+  { img: "/icon-service/5.webp", id: "service.mental" },
+  { img: "/icon-service/6.webp", id: "service.dental" },
+  { img: "/icon-service/7.webp", id: "service.surgery" },
+  { img: "/icon-service/8.webp", id: "service.diabetes" },
+  { img: "/icon-service/9.webp", id: "service.healthtest" },
+  { img: "/icon-service/10.webp", id: "service.nearby" },
+];
+
+export default services;
