+++ src/containers/HomePage/Data/specialtyData.js
@@ -1,0 +1,13 @@
+const specialties = [
+  { img: "/specialty-icon/1.webp", text: "specialty.musculoskeletal" },
+  { img: "/specialty-icon/2.webp", text: "specialty.neurology" },
+  { img: "/specialty-icon/3.webp", text: "specialty.digestive" },
+  { img: "/specialty-icon/4.webp", text: "specialty.cardiology" },
+  { img: "/specialty-icon/5.webp", text: "specialty.ent" },
+  { img: "/specialty-icon/6.webp", text: "specialty.spine" },
+  { img: "/specialty-icon/7.webp", text: "specialty.traditional_medicine" },
+  { img: "/specialty-icon/8.webp", text: "specialty.acupuncture" },
+  { img: "/specialty-icon/9.webp", text: "specialty.obstetrics_gynecology" },
+];
+
+export default specialties;
