+++ src/containers/HomePage/Section/Modal/ModalService.scss
@@ -1,0 +1,43 @@
+.modal-service {
+  position: fixed; /* Luôn cố định trên màn hình */
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%); /* Căn giữa chính xác */
+  width: 75%;
+  height: 80%;
+  padding: 20px;
+  background: white;
+  border-radius: 10px;
+  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.3);
+  z-index: 1000; /* Đảm bảo modal hiển thị trên cùng */
+
+  opacity: 1;
+
+  .specialty-links-container_md {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 1rem;
+    align-items: center;
+    width: 100%;
+
+    .specialty-container_md {
+      flex: 1 1 calc(18%);
+      max-width: calc(18%);
+
+      .specialty_option {
+        width: 100%;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        background: #f8f8f8;
+        border: 1px solid black;
+        padding: 15px;
+
+        img {
+          width: 100%;
+          height: 100%;
+        }
+      }
+    }
+  }
+}
